refacto multiple files + ext api call
